Print out size of removed trash.
t can run for quite a while if there is a lot of old trash to be removed
(i.e., on the order of gigabytes), so now a message is printed so the
user knows what's going on.

Since old trash is only removed at most once daily, this should not
cause overly verbose message spam.